Add RakutenStockAPI module and move data models
