+++ Sources/RakutenStockAPI/SymbolGroup.swift
